Buying product, other debuging.

+++ src/main/java/marketPlace/services/ProductServiceDbImplementation.java
@@ -8,9 +8,9 @@ import marketPlace.services.domain.ProductDomain;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Synchronization;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
 @Service
@@ -30,7 +30,7 @@ public class ProductServiceDbImplementation implements ProductService {
     }
 
     public String saveProduct(ProductModel productModel) {
-        if (!sellerServiceDbImplementation.existingSeller(productModel.getSellerId()))
+        if (sellerServiceDbImplementation.sellerIsNotExist(productModel.getSellerId()))
             throw new IllegalArgumentException("There is none seller with this sellerId");
 
         ProductDomain productDomain = productMapper.productModelToProductDomain(productModel);
@@ -41,6 +41,9 @@ public class ProductServiceDbImplementation implements ProductService {
     }
 
     public ProductModel getProductById(int productId) {
+        if (productIsNotExist(productId))
+            throw new IllegalArgumentException("There is none product with this productId");
+
         Product starter = productRepository.findById(productId).get();
         viewProduct(starter);
         ProductDomain productDomain = productMapper.productToProductDomain(starter);
@@ -57,17 +60,18 @@ public class ProductServiceDbImplementation implements ProductService {
 
     @Override
     public String deleteProduct(int productId) {
+        if (productIsNotExist(productId))
+            throw new IllegalArgumentException("There is none product with this productId");
         productRepository.deleteById(productId);
         return "Successful delete";
     }
 
     @Override
     public String modifyProduct(int productId, ProductModel productModel) {
-        if (!sellerServiceDbImplementation.existingSeller(productModel.getSellerId()))
-            throw new IllegalArgumentException("There is none seller with this sellerId");
+        if (productIsNotExist(productId))
+            throw new IllegalArgumentException("There is none product with this productId");
 
         Product starterProduct = productRepository.findById(productId).get();
-        viewProduct(starterProduct);
 
         ProductDomain oldProductDomain = productMapper.productToProductDomain(starterProduct);
 
@@ -79,25 +83,43 @@ public class ProductServiceDbImplementation implements ProductService {
         if (newProductDomain.getDescription() != null && !newProductDomain.getDescription().equals(oldProductDomain.getDescription()))
             oldProductDomain.setDescription(newProductDomain.getDescription());
 
-        if (newProductDomain.getPrice() < 0 && newProductDomain.getPrice() != oldProductDomain.getPrice())
+        if (newProductDomain.getPrice() > 0 && newProductDomain.getPrice() != oldProductDomain.getPrice())
             oldProductDomain.setPrice(newProductDomain.getPrice());
 
-        if (newProductDomain.getStock() < 0 && newProductDomain.getStock() != oldProductDomain.getStock())
+        if (newProductDomain.getStock() > 0 && newProductDomain.getStock() != oldProductDomain.getStock())
             oldProductDomain.setStock(newProductDomain.getStock());
 
         if (newProductDomain.getCategory() != null && !newProductDomain.getCategory().equals(oldProductDomain.getCategory()))
             oldProductDomain.setCategory(newProductDomain.getCategory());
 
-        if (newProductDomain.getSeller() != null && !newProductDomain.getSeller().equals(oldProductDomain.getSeller()))
-            oldProductDomain.setSeller(newProductDomain.getSeller());
-
         productRepository.save(productMapper.productDomainToProduct(oldProductDomain));
         return String.format("Successful modified product with %s id ", productId);
     }
 
-    private void viewProduct(Product product){
+    @Override
+    public synchronized String buyingProductById(int productId, int amount) {
+        if (productIsNotExist(productId))
+            throw new IllegalArgumentException("There is none product with this productId");
+        Product product = productRepository.findById(productId).get();
+        if (product.getStock() >= amount) {
+            product.decreaseStock(amount);
+        } else {
+            throw new IllegalArgumentException("There is none product in stock");
+        }
+        product.addNumberOfSales(amount);
+        productRepository.save(product);
+
+        return "Product was succesfully bougth";
+    }
+
+    // Helper methods
+    private void viewProduct(Product product) {
         int newValue = product.getNumberOfViewed() + 1;
         product.setNumberOfViewed(newValue);
         productRepository.save(product);
+    }
+
+    public boolean productIsNotExist(int productId) {
+        return !productRepository.existsById(productId);
     }
 }
